pyproject.toml
[project]
name = "summarizefromlinks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
